+++ app/View/Components/Form/DateInput.php
@@ -25,7 +25,7 @@ class DateInput extends Component
         $this->placeholder = $placeholder ?? $this->label;
         return <<<'blade'
             <x-form.wrapper>
-                <label for="{{$id}}">{{$label}} @if($required) <x-form.required-label /> @endif</label>
+                <label class="fw-bold" for="{{$id}}">{{$label}} @if($required) <x-form.required-label /> @endif</label>
                 <input class="form-control" type="date" name="{{$id}}" id="{{$id}}" placeholder="{{$placeholder}}" value="{{ old($id) }}">
                 <x-form.input-error-message id="{{$id}}" />
             </x-form.wrapper>