+++ app/View/Components/Form/MultiFileInput.php
@@ -11,7 +11,7 @@ class MultiFileInput extends Component
      *
      * @return void
      */
-    public function __construct(public $id, public $label, public $accept='image/*,.pdf', public $required=true, public $helperText="You may upload more than one file here. Please upload valid images or PDF files. Size of image should not be more than 2MB."){}
+    public function __construct(public $id, public $label, public $accept = 'image/*,.pdf', public $required = true, public $helperText = "You may upload more than one file here. Please upload valid images or PDF files. Size of image should not be more than 2MB."){}
 
     /**
      * Get the view / contents that represent the component.
@@ -21,14 +21,19 @@ class MultiFileInput extends Component
     public function render()
     {
         return <<<'blade'
-            <x-form.wrapper>
-                <label for="{{ $id }}">{{ $label }} @if($required) <x-form.required-label /> @endif</label>
-                <input type="file" class="form-control-file" accept="{{ $accept }}" name="{{ $id }}[]" id="{{ $id }}" multiple>
-                <small class="d-block form-text text-muted">{{$helperText}}</small>
-                <x-form.input-error-message id="{{$id}}" />
-            </x-form.wrapper>
+        <x-form.wrapper>
+            <div class="row">
+                <div class="col-12">
+                    <label for="{{$id}}" class="form-label fw-bold">{{$label}} @if($required) <x-form.required-label /> @endif</label>
+                </div>
+
+                <div class="col-md-6">
+                    <input type="file" class="form-control" accept="{{$accept}}" name="{{$id}}[]" id="{{$id}}" multiple>
+                    <small class="d-block form-text text-muted">{{$helperText}}</small>
+                </div>
+            </div>
+            <x-form.input-error-message id="{{$id}}" />
+        </x-form.wrapper>
         blade;
     }
 }
-
-