+++ app/View/Components/Form/NumberInput.php
@@ -26,7 +26,7 @@ class NumberInput extends Component
         return <<<'blade'
             <x-form.wrapper>
                 <label for="{{$id}}">{{$label}}@if($required) <x-form.required-label /> @endif</label>
-                <input class="form-control" type="number" name="{{$id}}" id="{{$id}}" placeholder="{{$placeholder}} value="{{ old($id) }}"">
+                <input class="form-control" type="number" name="{{$id}}" id="{{$id}}" placeholder="{{$placeholder}}" value="{{ old($id) }}">
                 <x-form.input-error-message id="{{$id}}" />
             </x-form.wrapper>
         blade;