+++ app/View/Components/Form/Radio.php
@@ -34,16 +34,30 @@ class Radio extends Component
     {
         return <<<'blade'
         <x-form.wrapper>
-            <label class="form-label d-block" for="{{$id}}">{{$label}} @if($required) <x-form.required-label /> @endif</label>
-            <div class="form-check">
-                @foreach($options as $option)
-                <input class="form-check-input" type="radio" name="{{$id}}" id="{{$id}}" value="{{$option[1]}}" @if(old($id) == $option[1]) checked @endif>
-                <label class="form-check-label mr-1" for="{{$id}}">{{$option[0]}}</label>
-                <br>
-                @endforeach
+
+            <div class= "row">
+                <div class="col-12">
+                    <label for="{{$id}}" class="fw-bold">{{$label}} @if($required) <x-form.required-label /> @endif</label>
+                </div>
+
+                <div class="col-md-6 mb-4">
+                    <div class="form-check">
+                        @foreach($options as $option)
+                        <input class="fw-bold form-check-input" type="radio" name="{{$id}}" id="{{$id}}"x-model="{{$id}}" value="{{$option[1]}}" @if(old($id) == $option[1]) checked @endif>
+                        <label class="form-check-label mr-1" for="{{$id}}">{{$option[0]}}</label>
+                        <br>
+                        @endforeach
+                    </div>
+                    <x-form.input-error-message id="{{$id}}" />
+                </div>
             </div>
-            <x-form.input-error-message id="{{$id}}" />
         </x-form.wrapper>
+        <style>
+        .form-check-input[type=radio] {
+            border-radius: 50%;
+            border-color: black;
+        }
+        </style>
         blade;
     }
 }