+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.2 - 26 MARCH 2025
+### Added
+- Updated buttons to Bootstrap styling
+- Added Alpine.js logic for conditional rendering based on radio input
+- Updated questions based on requirements
+- Updated layout to Bootstrap Row/Column Grid layout
+
 ## 1.0.1 - 24 FEBRUARY 2025
 ### Added
 - Added database seeders for areas, banks, hloes, and incomes